+++ src/DatalistSource.js
@@ -1,5 +1,7 @@
 import d from 'd_js';
 import Source from './Source.js';
+import Suggestion from './Suggestion.js';
+import Group from './Group.js';
 
 export default class DatalistSource extends Source {
     constructor(element, settings = {}) {
@@ -11,7 +13,47 @@ export default class DatalistSource extends Source {
         if (!settings.parent) {
             settings.parent = listElement.parentElement;
         }
-        super(getAvailableOptions(listElement), settings);
+
+        super(settings);
+
+        this.load(getAvailableOptions(listElement));
+    }
+
+    match(suggestion, query) {
+        return (
+            suggestion.label.toLowerCase().indexOf(query) !== -1 ||
+            suggestion.value.toLowerCase().indexOf(query) !== -1
+        );
+    }
+
+    refresh(query) {
+        this.element.innerHTML = '';
+        this.result = [];
+        this.current = 0;
+
+        if (!query) {
+            return this.close();
+        }
+
+        query = query.toLowerCase();
+
+        this.each((suggestion, parent) => {
+            suggestion.unselect();
+
+            if (this.match(suggestion, query)) {
+                parent.element.appendChild(suggestion.element);
+                this.result.push(suggestion);
+            } else if (suggestion.element.parentElement === parent.element) {
+                parent.element.removeChild(suggestion.element);
+            }
+        });
+
+        if (this.result.length) {
+            this.selectFirst();
+            this.open();
+        } else {
+            this.close();
+        }
     }
 }