+++ src/Suggestion.js
@@ -1,46 +1,25 @@
 import d from 'd_js';
 
 export default class Suggestion {
-    constructor(data, settings) {
+    constructor(data, settings = {}) {
         this.data = data;
         this.value = settings.value ? data[settings.value] : data.value;
         this.label = settings.label
             ? data[settings.label]
             : data.label || this.value;
 
-        this.search = settings.search || data.search;
-        this.render = settings.render;
-    }
-
-    set search(search) {
-        if (typeof search === 'function') {
-            this.searchData = search(this);
-        } else if (search) {
-            this.searchData = search;
-        } else {
-            this.searchData = `${this.data.label || ''} ${this
-                .value}`.toLowerCase();
-        }
-    }
-
-    set render(render) {
-        if (typeof render === 'function') {
-            this.element = d.parse(`<li>${render(this)}</li>`);
+        //Render
+        if (typeof settings.render === 'function') {
+            this.element = d.parse(`<li>${settings.render(this)}</li>`);
         } else {
             this.element = d.parse(`<li>${this.data.label || this.value}</li>`);
         }
     }
 
-    match(query) {
-        return query && this.searchData.indexOf(query.toLowerCase()) !== -1;
-    }
-
     refresh(parent, query, selected) {
         if (this.match(query)) {
-            if (this.element.parentElement !== parent) {
-                parent.appendChild(this.element);
-                this.unselect();
-            }
+            parent.appendChild(this.element);
+            this.unselect();
             selected.push(this);
         } else {
             if (this.element.parentElement === parent) {
@@ -49,8 +28,23 @@ export default class Suggestion {
         }
     }
 
-    select() {
+    detach() {
+        if (this.element.parentElement) {
+            this.element.parentElement.removeChild(this.element);
+        }
+    }
+
+    select(parent) {
+        const rect = this.element.getBoundingClientRect();
+        const parentRect = parent.getBoundingClientRect();
+
         this.element.classList.add('is-selected');
+
+        if (parentRect.top - rect.top > 0) {
+            this.element.scrollIntoView(true);
+        } else if (parentRect.bottom < rect.bottom) {
+            this.element.scrollIntoView(false);
+        }
     }
 
     unselect() {