+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.2.0 - UNRELEASED
+
+### Added
+
+* Auto-scroll the suggestions list if the currently selected element is not visible
+
+### Changed
+
+* Do not filter the suggestions in ajaxSource, it's supossed to be filtered by the server. 
+
+### Fixed
+
+* Remove the `is-selected` class to the suggestions on refresh.
+
+
 ## 0.1.0 - 2017-06-25
 
 First version with basic features